python/b34/pyproject.toml
[project]
name = "b34"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-ninja>=1.4.3",
    "requests>=2.32.5",
]
